Added basic poll response processing for the CM11A.
The CM11A can send two polls at any time, and continues to send these polls until the PC sends a response.  These polls were getting in the way of sending commands to the CM11A.  This change overrides the read() method and sends a basic poll response to the CM11A if a poll is received while attempting to read.  It then reads another byte to pass back to the calling method.